+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -1791,6 +1791,10 @@ public class ServiceHost {
             return this;
         }
 
+        if (this.isAuthorizationEnabled() && post.getAuthorizationContext() == null) {
+            populateAuthorizationContext(post);
+        }
+
         processServiceStart(ProcessingStage.INITIALIZING, service, post, post.hasBody());
         return this;
     }
@@ -1939,8 +1943,11 @@ public class ServiceHost {
 
                     if (post.hasBody()) {
                         document = post.getBody(s.getStateType());
+                    } else {
+                        document = new ServiceDocument();
+                        document.documentSelfLink = s.getSelfLink();
                     }
-                    if (!authorizeServiceState(s, document, post)) {
+                    if (!isAuthorized(s, document, post)) {
                         post.fail(Operation.STATUS_CODE_FORBIDDEN);
                         return;
                     }
@@ -2240,7 +2247,7 @@ public class ServiceHost {
         // If either there is cached state, or the service is not indexed (meaning nothing
         // will be found in the index), subject this state to authorization.
         if (state != null || !isServiceIndexed(s)) {
-            if (!authorizeServiceState(s, state, op)) {
+            if (!isAuthorized(s, state, op)) {
                 op.fail(Operation.STATUS_CODE_FORBIDDEN);
                 return;
             }
@@ -2274,7 +2281,7 @@ public class ServiceHost {
                     }
 
                     ServiceDocument st = o.getBody(s.getStateType());
-                    if (!authorizeServiceState(s, st, op)) {
+                    if (!isAuthorized(s, st, op)) {
                         op.fail(Operation.STATUS_CODE_FORBIDDEN);
                         return;
                     }
@@ -2291,7 +2298,12 @@ public class ServiceHost {
         this.documentIndexService.handleRequest(loadGet);
     }
 
-    private boolean authorizeServiceState(Service service, ServiceDocument document, Operation op) {
+    /**
+     * Infrastructure use. Applies authorization policy on the supplied document and fails the
+     * operation if authorization fails
+     * @return True if request was authorized, false otherwise
+     */
+    public boolean isAuthorized(Service service, ServiceDocument document, Operation op) {
         // Authorization not enabled, so there is nothing to check
         if (!this.isAuthorizationEnabled()) {
             return true;
@@ -2546,30 +2558,31 @@ public class ServiceHost {
             return true;
         }
 
-        if (inboundOp.getAuthorizationContext() == null) {
-            populateAuthorizationContext(inboundOp);
-        }
-
         if (this.isAuthorizationEnabled()) {
+            if (inboundOp.getAuthorizationContext() == null) {
+                populateAuthorizationContext(inboundOp);
+            }
+
             if (this.authorizationService != null) {
                 inboundOp.nestCompletion(op -> {
-                    handleAuthorizedRequest(service, op);
+                    handleRequestWithAuthContext(service, op);
                 });
                 queueOrScheduleRequest(this.authorizationService, inboundOp);
                 return true;
             }
         }
 
-        return handleAuthorizedRequest(service, inboundOp);
+        handleRequestWithAuthContext(service, inboundOp);
+        return true;
     }
 
-    private boolean handleAuthorizedRequest(Service service, Operation inboundOp) {
+    private void handleRequestWithAuthContext(Service service, Operation inboundOp) {
         String path;
         if (service == null) {
             path = inboundOp.getUri().getPath();
             if (path == null) {
                 failRequestServiceNotFound(inboundOp);
-                return true;
+                return;
             }
             service = findService(path);
         } else {
@@ -2583,22 +2596,32 @@ public class ServiceHost {
         }
 
         if (queueRequestUntilServiceAvailable(inboundOp, service, path)) {
-            return true;
+            return;
         }
 
         if (queueOrForwardRequest(service, path, inboundOp)) {
-            return true;
+            return;
         }
 
         if (service == null) {
             failRequestServiceNotFound(inboundOp);
-            return true;
+            return;
         }
 
         traceOperation(inboundOp);
 
+        if (isAuthorizationEnabled()) {
+            final Service sFinal = service;
+            inboundOp.nestCompletion((o) -> {
+                // only nest completion for success, failure will be forward to original completion
+                queueOrScheduleRequest(sFinal, inboundOp);
+            });
+            service.authorizeRequest(inboundOp);
+            return;
+        }
+
         queueOrScheduleRequest(service, inboundOp);
-        return true;
+        return;
     }
 
     AuthorizationContext getAuthorizationContext(Operation op) {
@@ -3220,6 +3243,8 @@ public class ServiceHost {
             this.attachedServices.remove(o.getUri().getPath());
             latch.countDown();
         };
+
+        setAuthorizationContext(getSystemAuthorizationContext());
 
         // first shut down non core services: During their stop processing they
         // might still rely on core services
@@ -4650,7 +4675,10 @@ public class ServiceHost {
             // 2) Call the service's getDocumentTemplate() to allow the service author to modify it
             // We are calling a function inside a lock, which is bad practice. This is however
             // by contract a synchronous function that should be O(1). We also only call it once.
-            desc = s.getDocumentTemplate().documentDescription;
+            ServiceDocumentDescription augmentedDesc = s.getDocumentTemplate().documentDescription;
+            if (augmentedDesc != null) {
+                desc = augmentedDesc;
+            }
 
             // 3) Update the cached entry
             this.descriptionCache.put(serviceTypeName, desc);