+++ xenon-common/src/main/java/com/vmware/xenon/services/common/ReliableSubscriptionService.java
@@ -99,6 +99,11 @@ public class ReliableSubscriptionService extends StatelessService {
     }
 
     @Override
+    public void authorizeRequest(Operation op) {
+        op.complete();
+    }
+
+    @Override
     public void handleRequest(Operation op) {
         if (!op.isNotification()) {
             super.handleRequest(op);