+++ xenon-common/src/main/java/com/vmware/xenon/services/common/authn/BasicAuthenticationService.java
@@ -40,6 +40,12 @@ public class BasicAuthenticationService extends StatelessService {
     private static final String BASIC_AUTH_SEPERATOR = " ";
     private static final String BASIC_AUTH_USER_SEPERATOR = ":";
 
+    @Override
+    public void authorizeRequest(Operation op) {
+        op.complete();
+    }
+
+    @Override
     public void handlePost(Operation op) {
         AuthenticationRequestType requestType = op.getBody(AuthenticationRequest.class).requestType;
         // default to login for backward compatibility